+++ pages/interests.js
@@ -1,0 +1,7 @@
+import Categories from '../components/categories/Categories.jsx';
+
+export default function InterestsPage() {
+  return (
+    <Categories/>
+  )
+}
